Extract flash/dry-run/diagnostics workers into gui_flash (slice 3a, move-only)
Move the serial operation workers and their plumbing out of FlasherFrame into
a new FlashController (gui_flash.py), following the gui_handset / gui_download
extraction idiom: the controller takes `frame` and reaches back through it for
everything only the frame owns; the frame keeps thin same-named delegators so
every existing call site keeps working unchanged.

Moved into FlashController:
  - on_flash + flash_thread (single KDH) / flash_thread_btf (RT-950 Pro) /
    batch_flash_thread + prompt_continue_batch
  - on_dry_run / dryrun_thread, on_diag / diag_thread
  - offer_test_report (with nag suppression) / offer_firmware_cleanup
  - the worker plumbing log_msg / set_progress / set_buttons

Stays on the frame as shared serial-error helpers (HandsetController's probe
thread also calls _log_dialout_hint): _is_permission_denied / _log_dialout_hint.
The selection model (_get_selected_radio / _driver_for /
_get_firmware_url_and_version) also stays frame-level; the controller reads it
through frame. Every worker->widget write still goes through wx.CallAfter.

This is a behavior-preserving move: the workers are relocated verbatim (only
self.X -> frame.X / self.plumbing rebinding). gui_main.py drops from 1826 to
1188 lines. TestOfferTestReportSuppression is repointed from
FlasherFrame._offer_test_report to gui_flash.FlashController.offer_test_report
(bound onto a controller stub with a fake frame), mirroring the earlier
"Point TestHandsetStatusConstants at gui_handset" repoint.

Slice 3 of the decompose-gui-main-workers change is gated by the manual
hardware checklist before merge; commit 3b converges the hand-rolled single-
flash / dry-run / diagnostics serial paths onto the driver functions.

python3 tests.py: 224 passed, 1 skipped.
